chore: combine test
