+++ libs/collab-stream/src/model.rs
@@ -1,6 +1,6 @@
 use collab_entity::CollabType;
 use std::collections::BTreeMap;
-use std::fmt::Display;
+use std::fmt::{Display, Formatter};
 use std::ops::Deref;
 use std::str::FromStr;
 
@@ -258,6 +258,25 @@ pub enum CollabControlEvent {
   Close {
     object_id: String,
   },
+}
+
+impl Display for CollabControlEvent {
+  fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+    match self {
+      CollabControlEvent::Open {
+        workspace_id: _,
+        object_id,
+        collab_type,
+        doc_state: _,
+      } => f.write_fmt(format_args!(
+        "Open collab: object_id:{}|collab_type:{:?}",
+        object_id, collab_type,
+      )),
+      CollabControlEvent::Close { object_id } => {
+        f.write_fmt(format_args!("Close collab: object_id:{}", object_id))
+      },
+    }
+  }
 }
 
 impl CollabControlEvent {
