+++ services/appflowy-history/src/core/manager.rs
@@ -104,7 +104,7 @@ async fn handle_control_event(
   handles: &Arc<DashMap<String, Arc<OpenCollabHandle>>>,
   pg_pool: &PgPool,
 ) {
-  trace!("Received control event: {:?}", event);
+  trace!("[History] received control event: {}", event);
   match event {
     CollabControlEvent::Open {
       workspace_id,
@@ -114,7 +114,7 @@ async fn handle_control_event(
     } => match handles.entry(object_id.clone()) {
       Entry::Occupied(_) => {},
       Entry::Vacant(entry) => {
-        trace!("Opening collab: {}", object_id);
+        trace!("[History] create collab: {}", object_id);
         match init_collab_handle(
           redis_stream,
           pg_pool,
@@ -136,7 +136,7 @@ async fn handle_control_event(
       },
     },
     CollabControlEvent::Close { object_id } => {
-      trace!("Close collab: {}", object_id);
+      trace!("[History] close collab: {}", object_id);
       if let Some(handle) = handles.get(&object_id) {
         if let Err(err) = handle.gen_history().await {
           error!(
