+++ tests/collab/mod.rs
@@ -1,11 +1,11 @@
 mod awareness_test;
 mod collab_curd_test;
 mod data_write_test;
-mod edit_permission;
 mod member_crud;
 mod missing_update_test;
 mod multi_devices_edit;
 mod pending_write_test;
+mod permission_test;
 mod single_device_edit;
 mod storage_test;
 pub mod util;
