+++ tests/collab_history/mod.rs
@@ -1,0 +1,1 @@
+mod document_history;
